paropt-service/requirements.txt
flask==1.0.2
globus-sdk
psycopg2-binary==2.8.1
pyopenssl
parsl[aws]==0.8.0
git+https://git@github.com/chaofengwu/paropt
redis>=3.2.1
# Using patched rq - a worker queue is broken in current release
# Update when next version is released (patch has been merged on github)
# rq>=1.0
git+https://git@github.com/macintoshpie/rq@patch-3

